+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 
-## 0.4.2-SNAPSHOT - pending
+## 0.4.3-SNAPSHOT - pending
+
+
+
+## 0.4.2 - 2023-09-29
+
+*	fix: handle case when cached key/value are obsolete after refactoring
+*	fix: handle case when @Cacheable method have no argument
+*	fix: handle generic return type for cache key
+*	build: update dependencies
 
 ## 0.4.1 - 2022-02-14